Added conftest
Separated fixtures and placed them in conftest.py
